requirements/base.txt
Flask==2.3.3
SQLAlchemy==2.0.20
python-dateutil==2.8.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
alembic==1.12.0
python-dotenv==1.0.0
Flask-Caching==2.0.2 
matlabengine==9.14.7